+++ build-mac.sh
@@ -5,7 +5,7 @@ cd $(dirname $(readlink -f $0))
 if test $(uname) = Linux; then
     rsync --exclude=release -av * bhj-mac:$(up .)
     rsync release bhj-mac:$(up .) -av -L --exclude=*/adb_usb_driver_smartisan
-    remote-cmd bhj-mac bash -c "set -x; cd $(up .); ./build-mac.sh"
+    remote-cmd bhj-mac bash -c "export DOING_T1WRENCH_RELEASE=$DOING_T1WRENCH_RELEASE; set -x; cd $(up .); ./build-mac.sh"
     rsync bhj-mac:$(up .)/T1Wrench.app ../T1Wrench-macos/ -av
     (
         cd ../T1Wrench-macos/
@@ -38,5 +38,9 @@ else
     rm T1Wrench.dmg -f
     macdeployqt T1Wrench.app -dmg -verbose=1 -executable=T1Wrench.app/Contents/MacOS/download
     mv ~/Qt5 ~/Qt5.bak
+    if test "$DOING_T1WRENCH_RELEASE"; then
+        exit
+    fi
+
     myscr bash -c 'ps-killall T1Wrench; of T1Wrench.app' || true
 fi
